requirements.txt
# Core dependencies
tkinter  # Usually included with Python
requests>=2.28.0
Pillow>=9.0.0

# System tray support
pystray>=0.19.0

# Optional: For real AI API integration
# openai>=1.0.0
# anthropic>=0.3.0

# Optional: For Microsoft authentication
# msal>=1.20.0
# requests-oauthlib>=1.3.0

# Development dependencies (optional)
# pytest>=7.0.0
# black>=22.0.0
# flake8>=5.0.0